lg-fastapi-exp/pyproject.toml
[project]
name = "langgraph-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "langchain-google-genai>=3.0.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
